Downgrade navigation library to 2.6.0, upgrade deps (#22)
* Bump deps, downgrade Jetpack Compose nav

Addresses an issue where navigation causes unexpected
slide/scale  animation when using compose navigation 2.7
and above. The solution is to downgrade to 2.6 and the problem
goes away.

* Restore current video position earlier in lifecycle

Sets the video's current time from the video player position listener
so that the time will be set during instantiation instead of on resume.
This avoids any judder when seeking to the video's time on orientation
change.