Fix tonemap and UV coordinates (#33)
Fixes the washed out AR video colors by calling
"inverseTonemapSRGB" on the video texture color. This
ensures that the tonemap is correctly set when the
video is recognized.

Additionally this commit removes the extra call to rotate
the final image texture. This is due to how Filament, the
3D library beneath ARCore, handles UV (XY) coordinates by
default

> When set to true (default value), the Y coordinate of
> UV attributes will be flipped when read by this
> material's vertex shader.

Relates to

- https://github.com/google/filament/issues/3067
- https://github.com/SceneView/sceneform-android/issues/437